Require a php file to check authentication
Require a php file to check authentication on xinha so that you have to be logged in to call the file. Which should only be possible if it is in the webroot which it is highly advised not to do. Always put the Jojo Core and mysite directories above the webroot it at all possible.